requirements.txt
selenium==4.17.2
selenium-wire==5.1.0
pytest==8.1.1
pytest-html==4.1.1
pytest-xdist==3.6.1
pytest-rerunfailures==12.0
python-dotenv==1.0.1
webdriver-manager==4.0.2
